initial work on eval_set tests
